Use tree to split into LV and payload.
The previous assumption that the parts are always stored in correct order was
(as suspected) not always correct. So we re-build the tree for ProtoVessel
and its ProtoPartSnapshots from parent pointers in the snapshots. I might have
used AttachNodeSnapshot for this, too. Not sure though. May need to rewrite
in the future.

For Vessel we use the parent and child pointers of each Part.

For both cases children are ordered according to their part name. That may need
to be improved in the future for asymmetric builds though.